i understand) why it doesnt work

+++ Weak3/code/Answer.java
@@ -19,8 +19,8 @@ public class Answer {
         a = b;
         b = temp;
     }
-    private static int factorial(int n){
-        int result = 1;
+    private static long factorial(int n){
+        long result = 1;
         if(n == 0 || n ==1){
             return result;
         }
